Updates to task views and user journey
