+++ src/router.js
@@ -91,13 +91,13 @@ import ExerciseTaskNew from '@/views/Exercise/Tasks/Task/New.vue';
 import ExerciseTaskStageOutcome  from '@/views/Exercise/Tasks/Task/StageOutcome/Overview.vue';
 import ExerciseTaskDataInitialised from '@/views/Exercise/Tasks/Task/Data/Initialised.vue';
 import ExerciseTaskDataActivated from '@/views/Exercise/Tasks/Task/Data/Activated.vue';
-import ExerciseTaskTestInitialised from '@/views/Exercise/Tasks/Task/TestInitialised.vue';
-import ExerciseTaskTestActivated from '@/views/Exercise/Tasks/Task/TestActivated.vue';
-import ExerciseTaskPanelsInitialised from '@/views/Exercise/Tasks/Task/PanelsInitialised.vue';
-import ExerciseTaskPanelsActivated from '@/views/Exercise/Tasks/Task/PanelsActivated.vue';
-import ExerciseTaskModerationInitialised from '@/views/Exercise/Tasks/Task/ModerationInitialised.vue';
-import ExerciseTaskModerationActivated from '@/views/Exercise/Tasks/Task/ModerationActivated.vue';
-import ExerciseTaskStatusChanges from '@/views/Exercise/Tasks/Task/StatusChanges.vue';
+import ExerciseTaskTestInitialised from '@/views/Exercise/Tasks/Task/Test/Initialised.vue';
+import ExerciseTaskTestActivated from '@/views/Exercise/Tasks/Task/Test/Activated.vue';
+import ExerciseTaskPanelsInitialised from '@/views/Exercise/Tasks/Task/Panels/Initialised.vue';
+import ExerciseTaskPanelsActivated from '@/views/Exercise/Tasks/Task/Panels/Activated.vue';
+import ExerciseTaskModerationInitialised from '@/views/Exercise/Tasks/Task/Moderation/Initialised.vue';
+import ExerciseTaskModerationActivated from '@/views/Exercise/Tasks/Task/Moderation/Activated.vue';
+import ExerciseTaskStatusChanges from '@/views/Exercise/Tasks/Task/StatusChanges/Index.vue';
 import ExerciseTaskFinalised from '@/views/Exercise/Tasks/Task/Finalised.vue';
 import ExerciseTaskFinalisedList from '@/views/Exercise/Tasks/Task/Finalised/List.vue';
 import ExerciseTaskFinalisedViewScore from '@/views/Exercise/Tasks/Task/Finalised/View.vue';
