+++ src/views/Exercise/Tasks/Task/Panels/Activated.vue
@@ -343,12 +343,12 @@
 </template>
 
 <script>
-import { beforeRouteEnter, btnNext } from './helper';
+import { beforeRouteEnter, btnNext } from '../helper';
 import Table from '@jac-uk/jac-kit/components/Table/Table.vue';
 import TableCell from '@jac-uk/jac-kit/components/Table/TableCell.vue';
 import TabsList from '@jac-uk/jac-kit/draftComponents/TabsList.vue';
 import ActionButton from '@jac-uk/jac-kit/draftComponents/ActionButton.vue';
-import { PANEL_TYPES, PANEL_STATUS } from './Panel/Constants';
+import { PANEL_TYPES, PANEL_STATUS } from '../Panel/Constants';
 import { CAPABILITIES, SELECTION_CATEGORIES, availableStatuses } from '@/helpers/exerciseHelper';
 import { getScoreSheetTotal, GRADE_VALUES } from '@/helpers/taskHelper';
 import { functions } from '@/firebase';
