+++ src/views/Exercise/Tasks/Task/Panels/Initialised.vue
@@ -194,13 +194,13 @@
 
 <script>
 // import { functions } from '@/firebase';
-import { beforeRouteEnter, btnNext } from './helper';
+import { beforeRouteEnter, btnNext } from '../helper';
 import Table from '@jac-uk/jac-kit/components/Table/Table.vue';
 import TableCell from '@jac-uk/jac-kit/components/Table/TableCell.vue';
 import TabsList from '@jac-uk/jac-kit/draftComponents/TabsList.vue';
 import Modal from '@jac-uk/jac-kit/components/Modal/Modal.vue';
 import TitleBar from '@/components/Page/TitleBar.vue';
-import SelectPanel from './Panel/components/SelectPanel.vue';
+import SelectPanel from '../Panel/components/SelectPanel.vue';
 import ActionButton from '@jac-uk/jac-kit/draftComponents/ActionButton.vue';
 import { functions } from '@/firebase';
